utils.jsを追加, それに伴った変更
